pyproject.toml
[project]
name = "radar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-driver>=0.2.9",
    "confluent-kafka>=2.11.1",
    "ipython>=9.5.0",
    "prometheus-client>=0.22.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
]
